LayoutTests/imported/w3c/web-platform-tests/css/css-transforms/animation/transform-composition-expected.txt


FAIL Compositing: property <transform> underlying [rotateX(100deg) rotateY(100deg)] from add [translate(10px, 20px)] to replace [rotateX(200deg) rotateY(200deg) translate(110px, 220px)] at (-0.5) should be [rotateX(50deg) rotateY(50deg) translate(-40px, -80px)] assert_equals: expected "matrix3d ( 0.64 , 0.59 , - 0.49 , 0 , 0 , 0.64 , 0.77 , 0 , 0.77 , - 0.49 , 0.41 , 0 , - 25.71 , - 74.9 , - 41.59 , 1 ) " but got "matrix3d ( 0.06 , - 0.94 , - 0.33 , 0 , 1 , 0.06 , 0.01 , 0 , 0.01 , - 0.33 , 0.94 , 0 , 66.68 , 126.93 , 55.3 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(100deg) rotateY(100deg)] from add [translate(10px, 20px)] to replace [rotateX(200deg) rotateY(200deg) translate(110px, 220px)] at (0) should be [rotateX(100deg) rotateY(100deg) translate(10px, 20px)] assert_equals: expected "matrix3d ( - 0.17 , 0.97 , 0.17 , 0 , 0 , - 0.17 , 0.98 , 0 , 0.98 , 0.17 , 0.03 , 0 , - 1.74 , 6.23 , 21.41 , 1 ) " but got "matrix ( 1 , 0 , 0 , 1 , 10 , 20 ) "
FAIL Compositing: property <transform> underlying [rotateX(100deg) rotateY(100deg)] from add [translate(10px, 20px)] to replace [rotateX(200deg) rotateY(200deg) translate(110px, 220px)] at (0.25) should be [rotateX(125deg) rotateY(125deg) translate(35px, 70px)] assert_equals: expected "matrix3d ( - 0.57 , 0.67 , 0.47 , 0 , 0 , - 0.57 , 0.82 , 0 , 0.82 , 0.47 , 0.33 , 0 , - 20.08 , - 16.66 , 73.79 , 1 ) " but got "matrix3d ( 0.73 , 0.68 , 0.07 , 0 , - 0.67 , 0.73 , - 0.17 , 0 , - 0.17 , 0.07 , 0.98 , 0 , - 18.34 , - 33.47 , - 27.65 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(100deg) rotateY(100deg)] from add [translate(10px, 20px)] to replace [rotateX(200deg) rotateY(200deg) translate(110px, 220px)] at (0.5) should be [rotateX(150deg) rotateY(150deg) translate(60px, 120px)] assert_equals: expected "matrix3d ( - 0.87 , 0.25 , 0.43 , 0 , 0 , - 0.87 , 0.5 , 0 , 0.5 , 0.43 , 0.75 , 0 , - 51.96 , - 88.92 , 85.98 , 1 ) " but got "matrix3d ( 0.06 , 1 , 0.01 , 0 , - 0.94 , 0.06 , - 0.33 , 0 , - 0.33 , 0.01 , 0.94 , 0 , - 46.68 , - 86.93 , - 55.3 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(100deg) rotateY(100deg)] from add [translate(10px, 20px)] to replace [rotateX(200deg) rotateY(200deg) translate(110px, 220px)] at (0.75) should be [rotateX(175deg) rotateY(175deg) translate(85px, 170px)] assert_equals: expected "matrix3d ( - 1 , 0.01 , 0.09 , 0 , 0 , - 1 , 0.09 , 0 , 0.09 , 0.09 , 0.99 , 0 , - 84.68 , - 168.71 , 22.2 , 1 ) " but got "matrix3d ( - 0.63 , 0.77 , - 0.15 , 0 , - 0.67 , - 0.63 , - 0.4 , 0 , - 0.4 , - 0.15 , 0.9 , 0 , - 75.02 , - 140.4 , - 82.95 , 1 ) "
PASS Compositing: property <transform> underlying [rotateX(100deg) rotateY(100deg)] from add [translate(10px, 20px)] to replace [rotateX(200deg) rotateY(200deg) translate(110px, 220px)] at (1) should be [rotateX(200deg) rotateY(200deg) translate(110px, 220px)] 
FAIL Compositing: property <transform> underlying [rotateX(100deg) rotateY(100deg)] from add [translate(10px, 20px)] to replace [rotateX(200deg) rotateY(200deg) translate(110px, 220px)] at (1.5) should be [rotateX(250deg) rotateY(250deg) translate(160px, 320px)] assert_equals: expected "matrix3d ( - 0.34 , 0.88 , - 0.32 , 0 , 0 , - 0.34 , - 0.94 , 0 , - 0.94 , - 0.32 , 0.12 , 0 , - 54.72 , 31.84 , - 352.12 , 1 ) " but got "matrix3d ( - 0.06 , - 0.93 , - 0.35 , 0 , 1 , - 0.06 , - 0.01 , 0 , - 0.01 , - 0.35 , 0.94 , 0 , - 160.05 , - 300.8 , - 165.9 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(45deg)] from add [none] to add [rotateY(360deg)] at (-0.5) should be [rotateX(45deg) rotateY(-180deg)] assert_equals: expected "matrix3d ( - 1 , 0 , 0 , 0 , 0 , 0.71 , 0.71 , 0 , 0 , 0.71 , - 0.71 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( - 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , - 1 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(45deg)] from add [none] to add [rotateY(360deg)] at (0) should be [rotateX(45deg) rotateY(0deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.71 , 0.71 , 0 , 0 , - 0.71 , 0.71 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix ( 1 , 0 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotateX(45deg)] from add [none] to add [rotateY(360deg)] at (0.25) should be [rotateX(45deg) rotateY(90deg)] assert_equals: expected "matrix3d ( 0 , 0.71 , - 0.71 , 0 , 0 , 0.71 , 0.71 , 0 , 1 , 0 , 0 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0 , 0 , - 1 , 0 , 0 , 1 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(45deg)] from add [none] to add [rotateY(360deg)] at (0.5) should be [rotateX(45deg) rotateY(180deg)] assert_equals: expected "matrix3d ( - 1 , 0 , 0 , 0 , 0 , 0.71 , 0.71 , 0 , 0 , 0.71 , - 0.71 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( - 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , - 1 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(45deg)] from add [none] to add [rotateY(360deg)] at (0.75) should be [rotateX(45deg) rotateY(270deg)] assert_equals: expected "matrix3d ( 0 , - 0.71 , 0.71 , 0 , 0 , 0.71 , 0.71 , 0 , - 1 , 0 , 0 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0 , 0 , 1 , 0 , 0 , 1 , 0 , 0 , - 1 , 0 , 0 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(45deg)] from add [none] to add [rotateY(360deg)] at (1) should be [rotateX(45deg) rotateY(360deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.71 , 0.71 , 0 , 0 , - 0.71 , 0.71 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(45deg)] from add [none] to add [rotateY(360deg)] at (1.5) should be [rotateX(45deg) rotateY(540deg)] assert_equals: expected "matrix3d ( - 1 , 0 , 0 , 0 , 0 , 0.71 , 0.71 , 0 , 0 , 0.71 , - 0.71 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( - 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , - 1 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(90deg)] from add [translate(100px, 100px)] to add [scale(2)] at (-0.5) should be [matrix3d(0.5, 0, 0, 0, 0, 1.11022e-16, 0.5, 0, 0, -1, 2.22045e-16, 0, 150, 9.18485e-15, 150, 1)] assert_equals: expected "matrix3d ( 0.5 , 0 , 0 , 0 , 0 , 0 , 0.5 , 0 , 0 , - 1 , 0 , 0 , 150 , 0 , 150 , 1 ) " but got "matrix ( 0.5 , 0 , 0 , 0.5 , 150 , 150 ) "
FAIL Compositing: property <transform> underlying [rotateX(90deg)] from add [translate(100px, 100px)] to add [scale(2)] at (0) should be [matrix3d(1, 0, 0, 0, 0, 6.12323e-17, 1, 0, 0, -1, 6.12323e-17, 0, 100, 6.12323e-15, 100, 1)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , - 1 , 0 , 0 , 100 , 0 , 100 , 1 ) " but got "matrix ( 1 , 0 , 0 , 1 , 100 , 100 ) "
FAIL Compositing: property <transform> underlying [rotateX(90deg)] from add [translate(100px, 100px)] to add [scale(2)] at (0.25) should be [matrix3d(1.25, 0, 0, 0, 0, 2.77556e-16, 1.25, 0, 0, -1, 2.22045e-16, 0, 75, 4.59243e-15, 75, 1)] assert_equals: expected "matrix3d ( 1.25 , 0 , 0 , 0 , 0 , 0 , 1.25 , 0 , 0 , - 1 , 0 , 0 , 75 , 0 , 75 , 1 ) " but got "matrix ( 1.25 , 0 , 0 , 1.25 , 75 , 75 ) "
FAIL Compositing: property <transform> underlying [rotateX(90deg)] from add [translate(100px, 100px)] to add [scale(2)] at (0.5) should be [matrix3d(1.5, 0, 0, 0, 0, 3.33067e-16, 1.5, 0, 0, -1, 2.22045e-16, 0, 50, 3.06162e-15, 50, 1)] assert_equals: expected "matrix3d ( 1.5 , 0 , 0 , 0 , 0 , 0 , 1.5 , 0 , 0 , - 1 , 0 , 0 , 50 , 0 , 50 , 1 ) " but got "matrix ( 1.5 , 0 , 0 , 1.5 , 50 , 50 ) "
FAIL Compositing: property <transform> underlying [rotateX(90deg)] from add [translate(100px, 100px)] to add [scale(2)] at (0.75) should be [matrix3d(1.75, 0, 0, 0, 0, 3.88578e-16, 1.75, 0, 0, -1, 2.22045e-16, 0, 25, 1.53081e-15, 25, 1)] assert_equals: expected "matrix3d ( 1.75 , 0 , 0 , 0 , 0 , 0 , 1.75 , 0 , 0 , - 1 , 0 , 0 , 25 , 0 , 25 , 1 ) " but got "matrix ( 1.75 , 0 , 0 , 1.75 , 25 , 25 ) "
FAIL Compositing: property <transform> underlying [rotateX(90deg)] from add [translate(100px, 100px)] to add [scale(2)] at (1) should be [matrix3d(2, 0, 0, 0, 0, 1.22465e-16, 2, 0, 0, -1, 6.12323e-17, 0, 0, 0, 0, 1)] assert_equals: expected "matrix3d ( 2 , 0 , 0 , 0 , 0 , 0 , 2 , 0 , 0 , - 1 , 0 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix ( 2 , 0 , 0 , 2 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotateX(90deg)] from add [translate(100px, 100px)] to add [scale(2)] at (1.5) should be [matrix3d(2.5, 0, 0, 0, 0, 5.55112e-16, 2.5, 0, 0, -1, 2.22045e-16, 0, -50, -3.06162e-15, -50, 1)] assert_equals: expected "matrix3d ( 2.5 , 0 , 0 , 0 , 0 , 0 , 2.5 , 0 , 0 , - 1 , 0 , 0 , - 50 , 0 , - 50 , 1 ) " but got "matrix ( 2.5 , 0 , 0 , 2.5 , - 50 , - 50 ) "
FAIL Compositing: property <transform> underlying [rotateX(45deg)] from add [scaleX(1)] to add [rotateY(360deg)] at (-0.5) should be [rotateX(45deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.71 , 0.71 , 0 , 0 , - 0.71 , 0.71 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(45deg)] from add [scaleX(1)] to add [rotateY(360deg)] at (0) should be [rotateX(45deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.71 , 0.71 , 0 , 0 , - 0.71 , 0.71 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix ( 1 , 0 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotateX(45deg)] from add [scaleX(1)] to add [rotateY(360deg)] at (0.25) should be [rotateX(45deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.71 , 0.71 , 0 , 0 , - 0.71 , 0.71 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(45deg)] from add [scaleX(1)] to add [rotateY(360deg)] at (0.5) should be [rotateX(45deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.71 , 0.71 , 0 , 0 , - 0.71 , 0.71 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(45deg)] from add [scaleX(1)] to add [rotateY(360deg)] at (0.75) should be [rotateX(45deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.71 , 0.71 , 0 , 0 , - 0.71 , 0.71 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(45deg)] from add [scaleX(1)] to add [rotateY(360deg)] at (1) should be [rotateX(45deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.71 , 0.71 , 0 , 0 , - 0.71 , 0.71 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(45deg)] from add [scaleX(1)] to add [rotateY(360deg)] at (1.5) should be [rotateX(45deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.71 , 0.71 , 0 , 0 , - 0.71 , 0.71 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 ) "
